backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.1
marshmallow==3.23.2
SQLAlchemy==2.0.36
pytest==8.3.4
PyJWT[crypto]==2.11.0
requests==2.32.5
